@@ -1,6 +1,6 @@
 {
   "name": "react-datagrid",
-  "version": "2.0.1",
+  "version": "2.1.0",
   "description": "React DataGrid",
   "main": "lib/index.js",
   "scripts": {
@@ -53,7 +53,8 @@
     "node-libs-browser": "^0.5.2",
     "parse-keys": "^1.3.0",
     "raw-loader": "^0.5.1",
-    "react-dom": "^0.14.2",
+    "react": "^15.1.0",
+    "react-dom": "^15.1.0",
     "react-hot-loader": "^1.2.7",
     "should": "^7.1.1",
     "sorty": "^1.2.0",
@@ -72,22 +73,20 @@
     "node-uuid": "^1.4.3",
     "normalize.css": "^3.0.3",
     "object-assign": "^4.0.1",
-    "react": "^0.14.0",
     "react-class": "^1.2.4",
-    "react-dom": "^0.14.0",
-    "react-load-mask": "^1.0.1",
-    "react-menus": "^2.0.4",
+    "react-load-mask": "^2.0.1",
+    "react-menus": "^2.0.5",
     "react-simple-toolbar": "^1.0.5",
     "react-style-normalizer": "^1.2.8",
-    "react-virtual-scroller": "^2.0.0",
+    "react-virtual-scroller": "^2.1.3",
     "region": "^2.1.2",
     "region-align": "^2.1.2",
     "ustring": "^1.4.1",
     "whatwg-fetch": "^0.10.0"
   },
   "peerDependencies": {
-    "react": ">=0.14.0",
-    "react-dom": ">=0.14.0"
+    "react": ">=0.14.0 || >=15.0.0",
+    "react-dom": ">=0.14.0 || >=15.0.0"
   },
   "author": "ZippyUI <contact@zippyui.com>",
   "contributors": [],
